refactor: profile related typing
