+++ shared/models/index.ts
@@ -11,10 +11,12 @@ export * from './maps'
 export * from './messages'
 export * from './moderation'
 export * from './notifications'
+export * from './patreon'
+export * from './patreonSettings'
+export * from './profile'
 export * from './research'
 export * from './storage'
 export * from './tags'
 export * from './questions'
 export * from './user'
 export * from './voteUseful'
-export * from './patreonSettings'
