+++ src/pages/UserSettings/content/fields/PatreonIntegration.tsx
@@ -4,7 +4,7 @@ import { PATREON_CLIENT_ID } from 'src/config/config'
 import { patreonService } from 'src/services/patreonService'
 import { Flex, Heading, Image, Text } from 'theme-ui'
 
-import type { IPatreonUser } from 'src/models/profile.model'
+import type { IPatreonUser } from 'oa-shared'
 
 export const HEADING = 'Patreon'
 export const SUCCESS_MESSAGE = 'Successfully linked Patreon account!'
