+++ src/routes/api.discussions.$sourceId.comments.ts
@@ -5,8 +5,7 @@ import { notificationsService } from 'src/services/notificationsService.server'
 import type { LoaderFunctionArgs } from '@remix-run/node'
 import type { Params } from '@remix-run/react'
 import type { User } from '@supabase/supabase-js'
-import type { DBCommentAuthor, Reply } from 'oa-shared'
-import type { DBProfile } from 'src/models/profile.model'
+import type { DBCommentAuthor, DBProfile,Reply  } from 'oa-shared'
 
 export async function loader({ params, request }: LoaderFunctionArgs) {
   if (!params.sourceId) {
